requirements.txt
GitPython
textual
